chore: refactor files tools
